+++ airbyte-cdk/python/airbyte_cdk/sources/streams/http/exceptions_async.py
